replace cdap with revised more vectorized cdap2
