--- activitysim/cdap/tests/data/cdap_3_person.csv
@@ -1,11 +1,0 @@
-Description,Expression,M,N,H
-All one,interaction == '111',1,0,0
-All two,interaction == '222',0,1,0
-All three,interaction == '333',0,0,1
-One + one + two,"interaction in ['112', '121', '211']",0,0,1
-One + one + three,"interaction in ['113', '131', '311']",0,0,1
-One + two + two,"interaction in ['122', '212', '221']",1,0,0
-One + two + three,"interaction in ['123', '132', '213', '231', '312', '321']",0,0,1
-One + three + three,"interaction in ['133', '313', '331']",0,1,0
-Two + two + three,"interaction in ['223', '232', '322']",0,0,1
-Two + three + three,"interaction in ['233', '323', '332']",1,0,0
