--- activitysim/cdap/tests/data/cdap_all_people.csv
@@ -1,7 +1,0 @@
-Description,Alternative,Value
-Three Person All M,"('M',) * 3",0
-Three Person All N,"('N',) * 3",0
-Three Person All H,"('H',) * 3",0
-Four Person All M,"('M',) * 4",0
-Four Person All N,"('N',) * 4",0
-Four Person All H,"('H',) * 4",0
