--- activitysim/cdap/tests/data/cdap_final_rules.csv
@@ -1,2 +1,0 @@
-Description,Expression,Alternative,Value
-M not allowed for Age 38 and Income 1900,(age == 38) & (income == 1900),M,0
