+++ activitysim/cdap/tests/data/cdap_interaction_coefficients.csv
@@ -1,0 +1,26 @@
+activity,interaction_ptypes,coefficient
+# 2-way interactions,,
+M,11,1
+N,22,1
+H,33,1
+M,12,1
+N,12,1
+H,13,1
+N,23,1
+# 3-way interactions,,
+H,124,0.9573
+# cdap_final_rules,,
+M,5,-999
+M,4,-999
+# cdap_all_people,,
+M,***,-0.0671
+N,***,-0.3653
+H,***,-1.1810
+M,****,-0.6104
+N,****,-1.3460
+H,****,-3.7330
+M,*****,-1.5280
+N,*****,-3.4530
+H,*****,-8.6210
+
+
