+++ activitysim/cdap/tests/data/people.csv
@@ -1,21 +1,21 @@
-id,household_id,ptype,age,income
-1,1,1,20,100
-2,2,3,21,200
-3,3,1,22,300
-4,3,1,23,400
-5,4,3,24,500
-6,4,2,25,600
-7,5,3,26,700
-8,5,2,27,800
-9,5,2,28,900
-10,6,1,29,1000
-11,6,3,30,1100
-12,6,1,31,1200
-13,7,1,32,1300
-14,7,3,33,1400
-15,7,2,34,1500
-16,7,2,35,1600
-17,8,1,36,1700
-18,8,3,37,1800
-19,8,2,38,1900
-20,8,1,39,2000
+id,household_id,ptype,age,PERSONS
+1,1,1,20,1
+2,2,3,21,1
+3,3,1,22,2
+4,3,1,23,2
+5,4,3,24,2
+6,4,2,25,2
+7,5,3,26,3
+8,5,2,27,3
+9,5,2,28,3
+10,6,1,29,3
+11,6,3,30,3
+12,6,1,31,3
+13,7,1,32,4
+14,7,3,33,4
+15,7,2,34,4
+16,7,2,35,4
+17,8,1,36,4
+18,8,3,37,4
+19,8,2,38,4
+20,8,1,39,4
