+++ activitysim/cdap/tests/test_cdap.py
@@ -25,150 +25,67 @@ def people(data_dir):
 
 
 @pytest.fixture(scope='module')
-def one_spec(data_dir):
+def cdap_indiv_and_hhsize1(data_dir):
     return read_model_spec(
-        os.path.join(data_dir, 'cdap_1_person.csv'))
+        os.path.join(data_dir, 'cdap_indiv_and_hhsize1.csv'))
 
 
 @pytest.fixture(scope='module')
-def two_spec(data_dir):
-    return read_model_spec(
-        os.path.join(data_dir, 'cdap_2_person.csv'))
-
-
-@pytest.fixture(scope='module')
-def three_spec(data_dir):
-    return read_model_spec(
-        os.path.join(data_dir, 'cdap_3_person.csv'))
-
-
-@pytest.fixture(scope='module')
-def final_rules(data_dir):
-    return read_model_spec(
-        os.path.join(data_dir, 'cdap_final_rules.csv'))
-
-
-@pytest.fixture
-def all_people(data_dir):
-    return read_model_spec(
-        os.path.join(data_dir, 'cdap_all_people.csv'),
-        expression_name='Alternative')
-
-
-@pytest.fixture(scope='module')
-def hh_id_col():
-    return 'household_id'
-
-
-@pytest.fixture(scope='module')
-def p_type_col():
-    return 'ptype'
+def cdap_interaction_coefficients(data_dir):
+    f = os.path.join(data_dir, 'cdap_interaction_coefficients.csv')
+    coefficients = pd.read_csv(f, comment='#')
+    coefficients = cdap.preprocess_interaction_coefficients(coefficients)
+    return coefficients
 
 
 @pytest.fixture(scope='module')
 def individual_utils(
-        people, hh_id_col, p_type_col, one_spec, two_spec, three_spec):
-    return cdap.individual_utilities(
-        people, hh_id_col, p_type_col, one_spec, two_spec, three_spec)
+        people, cdap_indiv_and_hhsize1):
+    return cdap.individual_utilities(people, cdap_indiv_and_hhsize1, locals_d=None)
 
 
-@pytest.fixture
-def hh_utils(individual_utils, people, hh_id_col):
-    hh_utils = cdap.initial_household_utilities(
-        individual_utils, people, hh_id_col)
-    return hh_utils
+# @pytest.fixture
+# def hh_utils(individual_utils, people, hh_id_col):
+#     hh_utils = cdap.initial_household_utilities(
+#         individual_utils, people, hh_id_col)
+#     return hh_utils
+#
+#
+# @pytest.fixture
+# def hh_choices(random_seed, hh_utils):
+#     return cdap.make_household_choices(hh_utils)
 
 
-@pytest.fixture
-def hh_choices(random_seed, hh_utils):
-    return cdap.make_household_choices(hh_utils)
+def test_bad_coefficients(data_dir):
+
+    f = os.path.join(data_dir, 'cdap_interaction_coefficients.csv')
+    coefficients = pd.read_csv(f, comment='#')
+
+    coefficients.loc[2, 'activity'] = 'AA'
+
+    with pytest.raises(RuntimeError) as excinfo:
+        coefficients = cdap.preprocess_interaction_coefficients(coefficients)
+    assert "Expect only M, N, or H" in str(excinfo.value)
 
 
-def test_make_interactions(people, hh_id_col, p_type_col):
-    expected_two = pd.DataFrame(
-        {'interaction': [
-            '11',  # household 3; person 3
-            '11',  # household 3; person 4
-            '32',  # household 4; person 5
-            '23',  # household 4; person 6
-            '32', '32',  # household 5; person 7
-            '23', '22',  # household 5; person 8
-            '23', '22',  # household 5; person 9
-            '13', '11',  # household 6; person 10
-            '31', '31',  # household 6; person 11
-            '11', '13',  # household 6; person 12
-            '13', '12', '12',  # household 7; person 13
-            '31', '32', '32',  # household 7; person 14
-            '21', '23', '22',  # household 7; person 15
-            '21', '23', '22',  # household 7; person 16
-            '13', '12', '11',  # household 8; person 17
-            '31', '32', '31',  # household 8; person 18
-            '21', '23', '21',  # household 8; person 19
-            '11', '13', '12'   # household 8; person 20
-        ]},
-        index=[
-            3, 4, 5, 6, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11, 12, 12,
-            13, 13, 13, 14, 14, 14, 15, 15, 15, 16, 16, 16,
-            17, 17, 17, 18, 18, 18, 19, 19, 19, 20, 20, 20
-        ])
+def test_assign_cdap_rank(people):
 
-    expected_three = pd.DataFrame(
-        {'interaction': [
-            '322', '322', '322',  # household 5; people 7, 8, 9
-            '131', '131', '131',  # household 6; people 10, 11, 12
-            '132', '132', '132',  # household 7; people 13, 14, 15
-            '132', '132', '132',  # household 7; people 13, 14, 16
-            '122', '122', '122',  # household 7; people 13, 15, 16
-            '322', '322', '322',  # household 7; people 14, 15, 16
-            '132', '132', '132',  # household 8; people 17, 18, 19
-            '131', '131', '131',  # household 8; people 17, 18, 20
-            '121', '121', '121',  # household 8; people 17, 19, 20
-            '321', '321', '321'   # household 8; people 18, 19, 20
-        ]},
-        index=[
-            7, 8, 9, 10, 11, 12,
-            13, 14, 15, 13, 14, 16, 13, 15, 16, 14, 15, 16,
-            17, 18, 19, 17, 18, 20, 17, 19, 20, 18, 19, 20
-        ])
+    cdap.assign_cdap_rank(people)
 
-    two, three = cdap.make_interactions(people, hh_id_col, p_type_col)
-
-    pdt.assert_frame_equal(two, expected_two)
-    pdt.assert_frame_equal(three, expected_three)
-
-
-def test_make_interactions_no_interactions(people, hh_id_col, p_type_col):
-    people = people.loc[[1, 2, 3]]
-
-    two, three = cdap.make_interactions(people, hh_id_col, p_type_col)
-
-    pdt.assert_frame_equal(
-        two, pd.DataFrame(columns=['interaction']), check_dtype=False)
-    pdt.assert_frame_equal(
-        three, pd.DataFrame(columns=['interaction']), check_dtype=False)
-
-
-def test_make_interactions_only_twos(people, hh_id_col, p_type_col):
-    people = people.loc[[1, 2, 3, 4, 5, 6]]
-
-    expected_two = pd.DataFrame(
-        {'interaction': [
-            '11',  # household 3; person 3
-            '11',  # household 3; person 4
-            '32',  # household 4; person 5
-            '23',  # household 4; person 6
-        ]},
-        index=[3, 4, 5, 6]
+    expected = pd.Series(
+        [1, 1, 1, 2, 2, 1, 3, 1, 2, 1, 3, 2, 1, 3, 2, 4, 1, 3, 4, 2],
+        index=people.index
     )
 
-    two, three = cdap.make_interactions(people, hh_id_col, p_type_col)
-
-    pdt.assert_frame_equal(two, expected_two)
-    pdt.assert_frame_equal(
-        three, pd.DataFrame(columns=['interaction']), check_dtype=False)
+    pdt.assert_series_equal(people['cdap_rank'], expected, check_dtype=False, check_names=False)
 
 
-def test_individual_utilities(people, one_spec, individual_utils):
+def test_individual_utilities(people, cdap_indiv_and_hhsize1):
+
+    cdap.assign_cdap_rank(people)
+    individual_utils = cdap.individual_utilities(people, cdap_indiv_and_hhsize1, locals_d=None)
+    individual_utils = individual_utils[['M', 'N', 'H']]
+
     expected = pd.DataFrame([
         [2, 0, 0],  # person 1
         [0, 0, 1],  # person 2
@@ -190,171 +107,32 @@ def test_individual_utilities(people, one_spec, individual_utils):
         [0, 0, 5],  # person 18
         [50, 0, 4],  # person 19
         [2, 0, 0]  # person 20
-        ], index=people.index, columns=one_spec.columns)
+        ], index=people.index, columns=cdap_indiv_and_hhsize1.columns)
 
     pdt.assert_frame_equal(
         individual_utils, expected, check_dtype=False, check_names=False)
 
 
-def test_initial_household_utilities(hh_utils):
-    alts = ['M', 'N', 'H']
-    one_alts = list(product(alts, repeat=1))
-    two_alts = list(product(alts, repeat=2))
-    three_alts = list(product(alts, repeat=3))
-    four_alts = list(product(alts, repeat=4))
+def test_build_cdap_spec_hhsize2(people, cdap_indiv_and_hhsize1, cdap_interaction_coefficients):
 
-    expected = {
-        1: pd.Series([2, 0, 0], index=one_alts),
-        2: pd.Series([0, 0, 1], index=one_alts),
-        3: pd.Series([6, 3, 3, 3, 0, 0, 3, 0, 0], index=two_alts),
-        4: pd.Series([1, 0, 0, 2, 1, 1, 1, 0, 0], index=two_alts),
-        5: pd.Series([
-            1, 1, 2, 3, 3, 4, 1, 1, 2,
-            0, 0, 1, 2, 2, 3, 0, 0, 1,
-            0, 0, 1, 2, 2, 3, 0, 0, 1,
-            ], index=three_alts),
-        6: pd.Series([
-            2, 2, 4, 2, 2, 4, 5, 5, 7,
-            0, 0, 2, 0, 0, 2, 3, 3, 5,
-            0, 0, 2, 0, 0, 2, 3, 3, 5,
-            ], index=three_alts),
-        7: pd.Series([
-            4, 8, 4, 8, 12, 8, 4, 8, 4,
-            3, 7, 3, 7, 11, 7, 3, 7, 3,
-            3, 7, 3, 7, 11, 7, 3, 7, 3,
-            1, 5, 1, 5, 9, 5, 1, 5, 1,
-            0, 4, 0, 4, 8, 4, 0, 4, 0,
-            0, 4, 0, 4, 8, 4, 0, 4, 0,
-            1, 5, 1, 5, 9, 5, 1, 5, 1,
-            0, 4, 0, 4, 8, 4, 0, 4, 0,
-            0, 4, 0, 4, 8, 4, 0, 4, 0
-            ], index=four_alts),
-        8: pd.Series([
-            52, 50, 50, 2, 0, 0, 6, 4, 4,
-            52, 50, 50, 2, 0, 0, 6, 4, 4,
-            57, 55, 55, 7, 5, 5, 11, 9, 9,
-            52, 50, 50, 2, 0, 0, 6, 4, 4,
-            52, 50, 50, 2, 0, 0, 6, 4, 4,
-            57, 55, 55, 7, 5, 5, 11, 9, 9,
-            56, 54, 54, 6, 4, 4, 10, 8, 8,
-            56, 54, 54, 6, 4, 4, 10, 8, 8,
-            61, 59, 59, 11, 9, 9, 15, 13, 13
-            ], index=four_alts)
-    }
+    hhsize = 2
 
-    assert list(hh_utils.keys()) == list(expected.keys())
-    for k in expected:
-        pdt.assert_series_equal(hh_utils[k], expected[k], check_dtype=False)
+    cdap.assign_cdap_rank(people)
+    indiv_utils = cdap.individual_utilities(people, cdap_indiv_and_hhsize1, locals_d=None)
 
+    choosers = cdap.hh_choosers(indiv_utils, hhsize=hhsize)
 
-def test_apply_final_rules(hh_utils, final_rules, people, hh_id_col):
-    expected = hh_utils.copy()
-    expected[8] = pd.Series([
-        0, 0, 0, 2, 0, 0, 6, 4, 4,
-        0, 0, 0, 2, 0, 0, 6, 4, 4,
-        0, 0, 0, 7, 5, 5, 11, 9, 9,
-        0, 0, 0, 2, 0, 0, 6, 4, 4,
-        0, 0, 0, 2, 0, 0, 6, 4, 4,
-        0, 0, 0, 7, 5, 5, 11, 9, 9,
-        0, 0, 0, 6, 4, 4, 10, 8, 8,
-        0, 0, 0, 6, 4, 4, 10, 8, 8,
-        0, 0, 0, 11, 9, 9, 15, 13, 13
-        ], index=expected[7].index)
+    spec = cdap.build_cdap_spec(cdap_interaction_coefficients, hhsize=hhsize)
 
-    cdap.apply_final_rules(hh_utils, people, hh_id_col, final_rules)
+    vars = cdap.eval_variables(spec.index, choosers)
 
-    for k in expected:
-        pdt.assert_series_equal(hh_utils[k], expected[k], check_dtype=False)
+    utils = vars.dot(spec).astype('float')
 
+    expected = pd.DataFrame([
+        [0, 3, 0, 3, 7, 3, 0, 3, 0],  # household 3
+        [0, 0, 1, 1, 1, 2, 0, 0, 2],  # household 4
+        ],
+        index=[3, 4],
+        columns=['HH', 'HM', 'HN', 'MH', 'MM', 'MN', 'NH', 'NM', 'NN']).astype('float')
 
-def test_apply_all_people(hh_utils, all_people):
-    all_people.at["('M',) * 3", 'Value'] = 300
-    all_people.at["('H',) * 4", 'Value'] = 500
-
-    expected = hh_utils.copy()
-    expected[5] = pd.Series([
-        301, 1, 2, 3, 3, 4, 1, 1, 2,
-        0, 0, 1, 2, 2, 3, 0, 0, 1,
-        0, 0, 1, 2, 2, 3, 0, 0, 1,
-        ], index=hh_utils[5].index)
-    expected[6] = pd.Series([
-        302, 2, 4, 2, 2, 4, 5, 5, 7,
-        0, 0, 2, 0, 0, 2, 3, 3, 5,
-        0, 0, 2, 0, 0, 2, 3, 3, 5,
-        ], index=hh_utils[6].index)
-    expected[7] = pd.Series([
-        4, 8, 4, 8, 12, 8, 4, 8, 4,
-        3, 7, 3, 7, 11, 7, 3, 7, 3,
-        3, 7, 3, 7, 11, 7, 3, 7, 3,
-        1, 5, 1, 5, 9, 5, 1, 5, 1,
-        0, 4, 0, 4, 8, 4, 0, 4, 0,
-        0, 4, 0, 4, 8, 4, 0, 4, 0,
-        1, 5, 1, 5, 9, 5, 1, 5, 1,
-        0, 4, 0, 4, 8, 4, 0, 4, 0,
-        0, 4, 0, 4, 8, 4, 0, 4, 500
-        ], index=hh_utils[7].index)
-    expected[8] = pd.Series([
-        52, 50, 50, 2, 0, 0, 6, 4, 4,
-        52, 50, 50, 2, 0, 0, 6, 4, 4,
-        57, 55, 55, 7, 5, 5, 11, 9, 9,
-        52, 50, 50, 2, 0, 0, 6, 4, 4,
-        52, 50, 50, 2, 0, 0, 6, 4, 4,
-        57, 55, 55, 7, 5, 5, 11, 9, 9,
-        56, 54, 54, 6, 4, 4, 10, 8, 8,
-        56, 54, 54, 6, 4, 4, 10, 8, 8,
-        61, 59, 59, 11, 9, 9, 15, 13, 513
-        ], index=hh_utils[8].index)
-
-    cdap.apply_all_people(hh_utils, all_people)
-
-    for k in expected:
-        pdt.assert_series_equal(hh_utils[k], expected[k], check_dtype=False)
-
-
-def test_make_household_choices(hh_choices):
-    expected = pd.Series([
-        ('M',),
-        ('H',),
-        ('M', 'M'),
-        ('N', 'N'),
-        ('M', 'N', 'H'),
-        ('M', 'H', 'H'),
-        ('M', 'M', 'N', 'N'),
-        ('H', 'H', 'M', 'N')],
-        index=range(1, 9))
-    pdt.assert_series_equal(hh_choices, expected)
-
-
-def test_household_choices_to_people(hh_choices, people):
-    people_choices = cdap.household_choices_to_people(hh_choices, people)
-    expected = pd.Series([
-        'M',
-        'H',
-        'M', 'M',
-        'N', 'N',
-        'M', 'N', 'H',
-        'M', 'H', 'H',
-        'M', 'M', 'N', 'N',
-        'H', 'H', 'M', 'N'],
-        index=people.index)
-    pdt.assert_series_equal(people_choices, expected)
-
-
-def test_run_cdap(
-        people, hh_id_col, p_type_col, one_spec, two_spec, three_spec,
-        final_rules, all_people, random_seed):
-
-    people_choices = cdap.run_cdap(
-        people, hh_id_col, p_type_col, one_spec, two_spec, three_spec,
-        final_rules, all_people)
-    expected = pd.Series([
-        'M',
-        'H',
-        'M', 'M',
-        'N', 'N',
-        'M', 'N', 'H',
-        'M', 'H', 'H',
-        'M', 'M', 'N', 'N',
-        'H', 'H', 'H', 'N'],
-        index=people.index)
-    pdt.assert_series_equal(people_choices, expected)
+    pdt.assert_frame_equal(utils, expected, check_names=False)
