+++ activitysim/defaults/models/__init__.py
@@ -12,4 +12,3 @@ import school_location
 import workplace_location
 import mode
 import cdap
-import xdap
