+++ activitysim/defaults/models/cdap.py
@@ -8,46 +8,62 @@ import pandas as pd
 
 from activitysim import activitysim as asim
 from activitysim import tracing
+
+from .util.misc import read_model_settings, get_model_constants
 from activitysim.cdap import cdap
 
-from .util.misc import read_model_settings
-from activitysim.cdap import xdap
+
+@orca.injectable()
+def cdap_settings(configs_dir):
+    """
+    canonical model settings file to permit definition of local constants for by
+    cdap_indiv_spec and cdap_fixed_relative_proportions
+    """
+
+    return read_model_settings(configs_dir, 'cdap.yaml')
 
 
 @orca.injectable()
-def cdap_1_person_spec(configs_dir):
-    f = os.path.join(configs_dir, 'cdap_1_person.csv')
+def cdap_indiv_spec(configs_dir):
+    """
+    spec to compute the activity utilities for each individual hh member
+    with no interactions with other household members taken into account
+    """
+
+    f = os.path.join(configs_dir, 'cdap_indiv_and_hhsize1.csv')
     return asim.read_model_spec(f).fillna(0)
 
 
 @orca.injectable()
-def cdap_2_person_spec(configs_dir):
-    f = os.path.join(configs_dir, 'cdap_2_person.csv')
-    return asim.read_model_spec(f).fillna(0)
+def cdap_interaction_coefficients(configs_dir):
+    """
+    Rules and coefficients for generating interaction specs for different household sizes
+    """
+    f = os.path.join(configs_dir, 'cdap_interaction_coefficients.csv')
+    return pd.read_csv(f, comment='#')
 
 
 @orca.injectable()
-def cdap_3_person_spec(configs_dir):
-    f = os.path.join(configs_dir, 'cdap_3_person.csv')
-    return asim.read_model_spec(f).fillna(0)
+def cdap_fixed_relative_proportions(configs_dir):
+    """
+    spec to compute/specify the relative proportions of each activity (M, N, H)
+    that should be used to choose activities for additional household members
+    not handled by CDAP
 
-
-@orca.injectable()
-def cdap_final_rules(configs_dir):
-    f = os.path.join(configs_dir, 'cdap_final_rules.csv')
-    return asim.read_model_spec(f).fillna(0)
-
-
-@orca.injectable()
-def cdap_all_people(configs_dir):
-    f = os.path.join(configs_dir, 'cdap_all_people.csv')
+    This spec is handled much like an activitysim logit utility spec,
+    EXCEPT that the values computed are relative proportions, not utilities
+    (i.e. values are not exponentiated before being normalized to probabilities summing to 1.0)
+    """
+    f = os.path.join(configs_dir, 'cdap_fixed_relative_proportions.csv')
     return asim.read_model_spec(f).fillna(0)
 
 
 @orca.step()
-def cdap_simulate(set_random_seed, persons_merged,
-                  cdap_1_person_spec, cdap_2_person_spec, cdap_3_person_spec,
-                  cdap_final_rules, cdap_all_people,
+def cdap_simulate(persons_merged,
+                  cdap_settings,
+                  cdap_indiv_spec,
+                  cdap_interaction_coefficients,
+                  cdap_fixed_relative_proportions,
                   hh_chunk_size, trace_hh_id):
     """
     CDAP stands for Coordinated Daily Activity Pattern, which is a choice of
@@ -59,32 +75,31 @@ def cdap_simulate(set_random_seed, persons_merged,
     simply applies those utilities using the simulation framework.
     """
 
-    choosers = persons_merged.to_frame()
+    persons_df = persons_merged.to_frame()
+
+    constants = get_model_constants(cdap_settings)
 
     tracing.info(__name__,
-                 "Running cdap_simulate with %d persons" % len(choosers.index))
+                 "Running cdap_simulate with %d persons" % len(persons_df.index))
 
-    choices = cdap.run_cdap(choosers,
-                            "household_id",
-                            "ptype",
-                            cdap_1_person_spec,
-                            cdap_2_person_spec,
-                            cdap_3_person_spec,
-                            cdap_final_rules,
-                            cdap_all_people,
-                            hh_chunk_size,
-                            trace_hh_id,
+    choices = cdap.run_cdap(persons=persons_df,
+                            cdap_indiv_spec=cdap_indiv_spec,
+                            cdap_interaction_coefficients=cdap_interaction_coefficients,
+                            cdap_fixed_relative_proportions=cdap_fixed_relative_proportions,
+                            locals_d=constants,
+                            chunk_size=hh_chunk_size,
+                            trace_hh_id=trace_hh_id,
                             trace_label='cdap')
 
     choices = choices.reindex(persons_merged.index)
 
-    tracing.print_summary('cdap_activity', choices, value_counts=True)
+    tracing.print_summary('cdap_activity', choices.cdap_activity, value_counts=True)
 
-    orca.add_column("persons", "cdap_activity", choices)
+    orca.add_column("persons", "cdap_activity", choices.cdap_activity)
+    orca.add_column("persons", "cdap_rank", choices.cdap_rank)
 
     if trace_hh_id:
-        trace_columns = ['cdap_activity']
         tracing.trace_df(orca.get_table('persons_merged').to_frame(),
                          label="cdap",
-                         columns=trace_columns,
+                         columns=['ptype', 'age', 'cdap_rank', 'cdap_activity'],
                          warn_if_empty=True)
