--- activitysim/defaults/test/configs/cdap_1_person.csv
@@ -1,55 +1,0 @@
-Description,Expression,Mandatory,NonMandatory,Home
-Full-time worker alternative-specific constants,ptype == 1,1.378734579,0.622662391,
-Part-time worker alternative-specific constants,ptype == 2,-0.718823738,0.636032467,
-University student alternative-specific constants,ptype == 3,2.353595176,0.609709846,
-Non-working adult alternative-specific constants,ptype == 4,-999,0.594645386,
-Retired alternative-specific constants,ptype == 5,-999,0.408202071,
-Driving-age child who is in school alternative-specific constants,ptype == 6,2.330918685,-0.599119112,
-Pre-driving-age child who is in school alternative-specific constants,ptype == 7,3.295863529,0.57142434,
-Pre-driving-age child who is too young for school alternative-specific constants,ptype == 8,1.052531189,-0.837567776,
-Pre-driving-age child who is too young for school interaction with age 0 to 1,(ptype == 8) & (age >= 0) & (age <= 1),-0.4515,,
-Pre-driving-age child who is too young for school interaction with age 4 to 5,(ptype == 8) & (age >= 4) & (age <= 5),0.6107,,
-Pre-driving-age child who is in school interaction with age 6 to 9,(ptype == 7) & (age >= 6) & (age <= 9),-0.2943,,
-Pre-driving-age child who is in school interaction with age 13 to 15,(ptype == 7) & (age >= 13) & (age <= 15),-0.7141,-0.672,
-Full-time worker interaction with age less than 40,(ptype == 1) & (age < 40),0.2091,,
-Retired interaction with age more than 80,(ptype == 5) & (age > 80),,,0.7666
-Full-time worker interaction with female gender,(ptype == 1) & (sex == 2),-0.1259,,
-Non-working adult interaction with female gender,(ptype == 4) & (sex == 2),-0.743,,
-Retired interaction with female,(ptype == 5) & (sex == 2),0.4769,,
-Non-working adult interaction with more cars than workers,(ptype == 4) & (auto_ownership > workers),0.6515,0.8168,
-Retired interaction with more cars than workers,(ptype == 5) & (auto_ownership > workers),2.992,1.056,
-Pre-driving-age child who is too young for school interaction with more cars than workers,(ptype == 8) & (auto_ownership > workers),,0.2991,
-Full-time worker interaction with fewer cars than workers,(ptype == 1) & (auto_ownership < workers),,,0.5039
-Non-working adult interaction with fewer cars than workers,(ptype == 4) & (auto_ownership < workers),,,0.8965
-Retired interaction with fewer cars than workers,(ptype == 5) & (auto_ownership < workers),,,0.5496
-Driving-age child who is in school interaction with fewer cars than workers,(ptype == 6) & (auto_ownership < workers),,,0.6475
-Pre-driving-age child who is in school interaction with fewer cars than workers,(ptype == 7) & (auto_ownership < workers),,,0.5862
-Pre-driving-age child who is too young for school interaction with fewer cars than workers,(ptype == 8) & (auto_ownership < workers),,,0.5061
-Full-time worker interaction with income less than $20k,(ptype == 1) & (income_in_thousands < 20),,,0.5313
-Retired interaction with income less than $20k,(ptype == 5) & (income_in_thousands < 20),,,0.533
-Part-time worker interaction with income less than $20k,(ptype == 2) & (income_in_thousands < 20),,,0.3232
-Part-time worker interaction with income between $50k and $100k,(ptype == 2) & (income_in_thousands >= 50) & (income_in_thousands <= 100),,,-0.4032
-Part-time worker interaction with income more than $100k,(ptype == 2) & (income_in_thousands < 100),,0.4207,-0.3534
-Non-working adult interaction with income between $50k and $100k,(ptype == 4) & (income_in_thousands >= 50) & (income_in_thousands <= 100),,,-0.5602
-Non-working adult interaction with income more than $100k,(ptype == 4) & (income_in_thousands < 100),,,-0.7188
-Driving-age child who is in school interaction with less than $20k,(ptype == 6) & (income_in_thousands < 20),,,1.307
-Driving-age child who is in school interaction income between $50k and $100k,(ptype == 6) & (income_in_thousands >= 50) & (income_in_thousands <= 100),,,-0.5031
-Driving-age child who is in school interaction with income more than $100k,(ptype == 6) & (income_in_thousands < 100),,,-2.046
-Pre-driving-age child who is too young for school interaction with income between $50k and $100k,(ptype == 8) & (income_in_thousands >= 50) & (income_in_thousands <= 100),,,-0.5708
-Pre-driving-age child who is too young for school interaction with income more than $100k,(ptype == 8) & (income_in_thousands < 100),,,-0.6186
-Full-time worker intraction with peak accessibility to all employment,(ptype == 1) * auPkTotal,0.1212,,
-Part-time worker interaction with peak accessibility to all employment,(ptype == 2) * auPkTotal,0.2004,,
-Non-working adult interaction with peak accessibility to all employment,(ptype == 4) * auPkTotal,0.2314,,
-Retired interaction with peak accessibility to all employment,(ptype == 5) * auPkTotal,0.2792,,
-Non-working adult interaction with off-peak accessibility to retail,(ptype == 4) * auOpRetail,,0.07207,
-Retired interaction with off-peak accessibility to retail,(ptype == 5) * auOpRetail,,0.07207,
-University student interaction with off-peak accessibility to retail,(ptype == 3) * auOpRetail,,0.07207,
-Driving-age child who is in school interaction with off-peak accessibility to retail,(ptype == 6) * auOpRetail,,0.08233,
-Pre-driving-age child who is in school interaction with off-peak accessibility to retail,(ptype == 7) * auOpRetail,,0.08233,
-Pre-driving-age child who is too young for school interaction with off-peak accessibility to retail,(ptype == 8) * auOpRetail,,0.08233,
-# Full-time worker interaction with usual work location is home,(ptype == 1) * usualWorkLocationIsHome,-1.758,,0.1813
-# Part-time worker interaction with usual work location is home,(ptype == 2) * usualWorkLocationIsHome,-1.758,,0.1813
-# Full-time worker interaction with no usual work location,(ptype == 1) * noUsualWorkLocation,-0.5935,,
-# Part-time worker interaction with no usual work location,(ptype == 2) * noUsualWorkLocation,-0.5935,,
-# Driving-age child who is in school interaction with no usual school location,(ptype == 6) * noUsualWorkLocation,-0.866,,
-# Pre-driving age child who is in school interaction with no usual school location,(ptype == 7) * noUsualWorkLocation,-0.866,,
