+++ activitysim/defaults/test/configs/cdap_interaction_coefficients.csv
@@ -1,0 +1,140 @@
+activity,interaction_ptypes,coefficient
+# 2-way interactions,,
+H,11,1.626
+H,12,0.7407
+H,13,1.183
+H,14,0.9436
+H,15,1.298
+H,16,2.064
+H,17,1.501
+H,18,0.9912
+H,22,0.8911
+H,23,1.642
+H,24,0.7057
+H,25,0.463
+H,26,3.057
+H,27,0.7685
+H,28,1.07
+H,33,1.018
+H,34,1.781
+H,35,0.4835
+H,36,1.546
+H,37,1.552
+H,38,1.34
+H,44,1.352
+H,45,1.209
+H,46,0.5243
+H,47,0.8112
+H,48,1.167
+H,55,1.407
+H,56,0.8632
+H,57,0.8632
+H,58,0.8632
+H,66,2.198
+H,67,0.977
+H,68,1.467
+H,77,2.8
+H,78,1.434
+H,88,1.378
+M,11,0.141
+M,12,0.08845
+M,13,0.4273
+M,16,0.3842
+M,17,0.2623
+M,18,0.5118
+M,22,1.135
+M,23,0.173
+M,26,1.103
+M,27,0.3079
+M,28,0.5074
+M,33,0.8726
+M,36,-0.0021
+M,37,0.2975
+M,38,0.2254
+M,66,0.4794
+M,67,0.5151
+M,68,0.5516
+M,77,0.9731
+M,78,0.5961
+M,88,1.651
+N,11,1.123
+N,12,0.4947
+N,13,0.5523
+N,14,0.02186
+N,15,0.3115
+N,16,0.4095
+N,17,0.6008
+N,18,0.751
+N,22,1.032
+N,23,0.3355
+N,24,0.7477
+N,25,0.09831
+N,26,0.495
+N,27,0.8984
+N,28,1.452
+N,33,1.054
+N,34,0.193
+N,35,0.4065
+N,36,1.62
+N,37,0.5165
+N,38,0.8973
+N,44,0.6984
+N,45,0.1864
+N,46,0.6801
+N,47,0.5646
+N,48,1.164
+N,55,0.7291
+N,56,0.2919
+N,57,0.2919
+N,58,0.2919
+N,66,1.512
+N,67,1.422
+N,68,1.273
+N,77,1.553
+N,78,0.6184
+N,88,0.8771
+# 3-way interactions,,
+H,124,0.9573
+H,122,0.9573
+H,144,0.9573
+H,126,0.2939
+H,146,0.2939
+H,222,0.9881
+H,224,0.9881
+H,244,0.9881
+H,226,0.4374
+H,246,0.4374
+H,446,0.4374
+H,266,0.4747
+H,466,0.4747
+M,111,0.3133
+M,112,0.3495
+M,114,0.3495
+M,666,-0.3906
+N,112,0.4637
+N,114,0.4637
+N,124,0.3491
+N,122,0.3491
+N,144,0.3491
+N,166,0.3553
+N,222,-1.386
+N,224,-1.386
+N,444,-1.386
+N,246,-0.8571
+N,226,-0.8571
+N,446,-0.8571
+# cdap_final_rules,,
+M,5,-999
+M,4,-999
+# cdap_all_people,,
+M,***,-0.0671
+N,***,-0.3653
+H,***,-1.1810
+M,****,-0.6104
+N,****,-1.3460
+H,****,-3.7330
+M,*****,-1.5280
+N,*****,-3.4530
+H,*****,-8.6210
+
+
