+++ example/configs/cdap.yaml
@@ -1,0 +1,9 @@
+CONSTANTS:
+  FULL: 1
+  PART: 2
+  UNIVERSITY: 3
+  NONWORK: 4
+  RETIRED: 5
+  DRIVING: 6
+  SCHOOL: 7
+  PRESCHOOL: 8
