--- sandbox/configs/cdap_2_person.csv
@@ -1,37 +1,0 @@
-Description,Expression,Mandatory,NonMandatory,Home
-Full-time worker interaction with full-time worker,interaction == '11',0.141,1.123,1.626
-Full-time worker interaction with part-time worker,"interaction in ['12',  '21']",0.08845,0.4947,0.7407
-Full-time worker interaction with university student,"interaction in ['13',  '31']",0.4273,0.5523,1.183
-Full-time worker interaction with non-working adult,"interaction in ['14',  '41']",,0.02186,0.9436
-Full-time worker interaction with retired,"interaction in ['15',  '51']",,0.3115,1.298
-Full-time worker interaction with driving-age child who is in school,"interaction in ['16',  '61']",0.3842,0.4095,2.064
-Full-time worker interaction with pre-driving-age child who is in school,"interaction in ['17',  '71']",0.2623,0.6008,1.501
-Full-time worker interaction with pre-driving-age child who is too young for school,"interaction in ['18',  '81']",0.5118,0.751,0.9912
-Part-time worker interaction with part-time worker,interaction == '22',1.135,1.032,0.8911
-Part-time worker interaction with university student,"interaction in ['23',  '32']",0.173,0.3355,1.642
-Part-time worker interaction with non-working adult,"interaction in ['24',  '42']",,0.7477,0.7057
-Part-time worker interaction with retired,"interaction in ['25',  '52']",,0.09831,0.463
-Part-time worker interaction with driving-age child who is in school,"interaction in ['26',  '62']",1.103,0.495,3.057
-Part-time worker interaction with pre-driving-age child who is in school,"interaction in ['27',  '72']",0.3079,0.8984,0.7685
-Part-time worker interaction with pre-driving-age child who is too young for school,"interaction in ['82',  '82']",0.5074,1.452,1.07
-University student interaction with university student,interaction == '33',0.8726,1.054,1.018
-University student interaction with non-working adult,"interaction in ['34',  '43']",,0.193,1.781
-University student interaction with retired,"interaction in ['35',  '53']",,0.4065,0.4835
-University student interaction with driving-age child who is in school,"interaction in ['36',  '63']",-0.0021,1.62,1.546
-University student interaction with pre-driving-age child who is in school,"interaction in ['37',  '73']",0.2975,0.5165,1.552
-University student interaction with pre-driving-age child who is too young for school,"interaction in ['38',  '83']",0.2254,0.8973,1.34
-Non-working adult interaction with non-working adult,interaction == '44',,0.6984,1.352
-Non-working adult interaction with retired,"interaction in ['45',  '54']",,0.1864,1.209
-Non-working adult interaction with driving-age child who is in school,"interaction in ['46',  '64']",,0.6801,0.5243
-Non-working adult interaction with pre-driving-age child who is in school,"interaction in ['47',  '74']",,0.5646,0.8112
-Non-working adult interaction with pre-driving-age child who is too young for school,"interaction in ['48',  '84']",,1.164,1.167
-Retired interaction with retired,interaction == '55',,0.7291,1.407
-Retired interaction with driving-age child who is in school,"interaction in ['56',  '65']",,0.2919,0.8632
-Retired interaction with pre-driving-age child who is in school,"interaction in ['57',  '75']",,0.2919,0.8632
-Retired interaction with pre-driving-age child who is too young for school,"interaction in ['58',  '85']",,0.2919,0.8632
-Driving-age child who is in school interaction with driving-age child who is in school,interaction == '66',0.4794,1.512,2.198
-Driving-age child who is in school interaction with pre-driving-age child who is in school,"interaction in ['67',  '76']",0.5151,1.422,0.977
-Driving-age child who is in school interaction with pre-driving-age child who is too young for school,"interaction in ['68',  '86']",0.5516,1.273,1.467
-Pre-driving-age child who is in school interaction with pre-Pre-driving-age child who is in school,interaction == '77',0.9731,1.553,2.8
-Pre-driving-age child who is in school interaction with pre-driving-age child who is too young for school,"interaction in ['78',  '87']",0.5961,0.6184,1.434
-Pre-driving-age child who is too young for school interaction with pre-driving-age child who is too young for school,interaction == '88',1.651,0.8771,1.378
