--- sandbox/configs/cdap_3_person.csv
@@ -1,10 +1,0 @@
-Description,Expression,Mandatory,NonMandatory,Home
-Full-time worker / full-time worker / full-time worker,interaction == '111',0.3133,,
-Full-time worker / full-time worker / part-time worker or non-working adult,"interaction in ['112', '121', '211', '114', '141', '411']",0.3495,0.4637,
-Full-time worker / part-time worker or non-working adult / part-time worker or non-working adult,"interaction in ['124', '142', '214', '412', '241', '421', '122', '144', '212', '414', '221', '441']",,0.3491,0.9573
-Full-time worker / part-time worker or non-working adult / pre-driving-age child,"interaction in ['162', '126', '216', '612', '261', '621']",,,0.2939
-Full-time worker / pre-driving-age child / pre-driving-age child,"interaction in ['166', '616', '661']",,0.3553,
-Part-time worker or non-working adult / part-time worker or non-working adult / part-time worker or non-working adult,"interaction in ['222', '224', '242', '422', '424', '442', '244', '444']",,-1.386,0.9881
-Part-time worker or non-working adult / part-time worker or non-working adult / pre-driving-age child,"interaction in ['624', '642', '264', '462', '246', '426', '622', '644', '262', '464', '226', '446']",,-0.8571,0.4374
-Part-time worker or non-working adult / pre-driving-age child / pre-driving-age child,"interaction in ['662', '626', '266', '664', '646', '466']",,,0.4747
-Pre-driving-age child / pre-driving-age child / pre-driving-age child,interaction == '666',-0.3906,,
